More tests plus cleanup
Added multiple sets of internal and external thread cross sections
so that I could test both slopes. Doing so allowed me to see and find
a new bug where the "top" of the external thread is not parallel to
the "bottom" slope of the internal thread. NO FIX YET.

More work to do