Threads is now passed the cutoff sizes directly
Previously the cutoffs were a divisor for pitch, this is how they are
specified by the european standards, but specifying the cutoff sizes
directly makes more sense.

Also, remove using Union[float], previously it was a Union of float and
None but I changed that awhile ago but didn't remove the Union.